requirements/base.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg==3.2.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7